Add autofill-survey to password management page
